+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [8.45] - 2026-04-18
+### Added
+- Added a new option to track and sync "Alive Time" in local storage.
+- Added a new top menu bar element to display a live updating relative Alive Time.
+- Added a mechanism to wipe local tracking if the player's life drops to 0%.
+- Added sync logic to the Living Area that gracefully updates your alive tracker.
+
 ## [8.44] - 2026-04-17
 ### Added
 - Added "Check for Updates" button to the Settings menu that fetches the latest release from GitHub to compare against current version.